executor/pyproject.toml
[tool.poetry]
name = "teddy-executor"
version = "0.1.0"
description = "A CLI tool to safely execute AI-generated YAML plans."
authors = ["AI Architect"]
license = "MIT"
readme = "README.md"
packages = [{ include = "teddy", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10"
typer = { version = "*", extras = ["all"] }
pyyaml = "*"
pyperclip = "*"
markdownify = "*"
playwright = "*"
httpx = "^0.28.1"
beautifulsoup4 = "^4.14.3"
requests = "^2.32.5"
ddgs = "^9.10.0"
pathspec = "^0.12.1"

[tool.poetry.group.dev.dependencies]
pytest = "*"
ruff = "*"
mypy = "*"
pre-commit = "*"
types-PyYAML = "*"
responses = "^0.25.8"
pytest-httpserver = "^1.1.3"
types-requests = "^2.32.4.20250913"

[tool.poetry.scripts]
teddy = "teddy.main:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
exclude = ["spikes/"]

[tool.mypy]
ignore_missing_imports = true
exclude = ["spikes/"]
